[pom] Add in override on compiler for when release flag is used
due to some bugs in compiler plugin, will need to skip the profile entirely to get it to work.  Doing so results in not needing this but if we need to research further, we will need this.  The downgrade here is for hibernate validator to older copy so it does not get 'hijrahDate' issue.