[ci] Update old todo comment on why dtd not tested
servlet-api 2.4 was originally in code base that included the web content files such as dtd, xsd, etc.  That is in implementations now such as tomcat not the api.  This code base did further include it on the build but I removed that long time back as unnecessary to test this code base has that consideration.  Keeping it todo as attempt to bring in tomcat api just fails which was indicated on my original attempts.